config| Moves mock objects into separate file
